Implement and separate interface for VQVAE; make image tokenizer generic
