+++ examples/train_colored_mnist_images.py
@@ -2,8 +2,9 @@ from typing import cast
 
 from torchsmith.datahub.colored_mnist import ColoredMNISTDataset
 from torchsmith.models.gpt2 import GPT2Decoder
-from torchsmith.tokenizers.mnist_tokenizer import VQVAEColoredMNISTImageTokenizer
+from torchsmith.tokenizers.mnist_tokenizer import VQVAEMNIST
 from torchsmith.tokenizers.mnist_tokenizer import generate_samples_colored_mnist_image
+from torchsmith.tokenizers.vqvae_tokenizer import VQVAEImageTokenizer
 from torchsmith.training.config import GPT2Config
 from torchsmith.training.config import TrainConfig
 from torchsmith.training.data import DataHandler
@@ -21,7 +22,7 @@ n_jobs = 12
 
 set_resource_limits(n_jobs=n_jobs, maximum_memory=26)
 
-tokenizer = VQVAEColoredMNISTImageTokenizer(batch_size=10000)
+tokenizer = VQVAEImageTokenizer(vqvae=VQVAEMNIST(), batch_size=10000)
 
 train_config = TrainConfig(num_epochs=100, batch_size=512, num_workers=n_jobs)
 transformer_config = GPT2Config(seq_len=ColoredMNISTDataset.SEQUENCE_LENGTH)
