+++ src/torchsmith/models/vae/base.py
@@ -31,3 +31,20 @@ class BaseVAE(ABC, torch.nn.Module):
     @abstractmethod
     def reconstruct(self, x: torch.Tensor) -> torch.Tensor:
         raise NotImplementedError()
+
+
+class BaseVQVAE(ABC, torch.nn.Module):
+    @abstractmethod
+    @torch.no_grad()
+    def encode_to_indices(self, x: torch.Tensor) -> torch.Tensor:
+        raise NotImplementedError()
+
+    @abstractmethod
+    @torch.no_grad()
+    def decode_from_indices(self, x: torch.Tensor) -> torch.Tensor:
+        raise NotImplementedError()
+
+    @property
+    @abstractmethod
+    def codebook_size(self) -> int:
+        raise NotImplementedError()
