+++ src/torchsmith/tokenizers/mnist_tokenizer.py
@@ -1,4 +1,3 @@
-from collections.abc import Iterator
 from typing import cast
 
 import numpy as np
@@ -9,124 +8,41 @@ from torchsmith.datahub.colored_mnist import ColoredMNISTWithTextDataset
 from torchsmith.datahub.colored_mnist import get_unique_tokens
 from torchsmith.models.external.colored_mnist import load_pretrain_vqvae
 from torchsmith.models.gpt2 import GPT2Decoder
+from torchsmith.models.vae.base import BaseVQVAE
 from torchsmith.tokenizers.base_tokenizer import BaseTokenizer
 from torchsmith.tokenizers.string_tokenizer import WordTokenizer
+from torchsmith.tokenizers.vqvae_tokenizer import VQVAEImageTokenizer
 from torchsmith.utils.dtypes import TokenType
 from torchsmith.utils.plotting import darken_fraction_of_image
 from torchsmith.utils.plotting import plot_images
 from torchsmith.utils.pytorch import get_device
-from torchsmith.utils.pyutils import batched
 
 device = get_device()
 
 
-class VQVAEColoredMNISTImageTokenizer(BaseTokenizer[TokenType]):
-    BOS = "<boi>"  # Beginning of image.
-    EOS = "<eoi>"  # End of image.
+class VQVAEMNIST(BaseVQVAE):
+    def __init__(self) -> None:
+        super().__init__()
+        self.vqvae = load_pretrain_vqvae().eval().to(device)
 
-    def __init__(
-        self,
-        *,
-        n_jobs: int = 1,
-        batch_size: int = 10000,
-        verbose: bool = False,
-        token_id_offset: int = 0,
-    ) -> None:
-        super().__init__(n_jobs=n_jobs, batch_size=batch_size, verbose=verbose)
-        self.device = get_device()
-        self.vqvae = load_pretrain_vqvae().eval().to(self.device)
-        self._id_to_token: dict[int, TokenType] = {
-            (index + token_id_offset): token
-            for index, token in enumerate(range(self.vqvae.n_embeddings))
-        }
-        self._id_to_token.update(
-            {max(self._id_to_token) + 1: self.BOS, max(self._id_to_token) + 2: self.EOS}
+    def encode_to_indices(self, x: torch.Tensor) -> torch.Tensor:
+        return cast(
+            torch.Tensor, self.vqvae.quantize(x.numpy()).reshape(x.shape[0], -1)
         )
-        self._token_to_id = {token: idx for idx, token in self._id_to_token.items()}
-        if self.verbose:
-            print(
-                f"Created a mapping for {len(self)} unique tokens: {self._token_to_id}"
-            )
 
-    @property
-    def token_to_id(self) -> dict[TokenType, int]:
-        return self._token_to_id
-
-    @property
-    def id_to_token(self) -> dict[int, TokenType]:
-        return self._id_to_token
-
-    def image_to_sequence(self, x: np.ndarray) -> np.ndarray:
+    def decode_from_indices(self, x: torch.Tensor) -> torch.Tensor:
         with torch.no_grad():
-            return (
-                cast(torch.Tensor, self.vqvae.quantize(x).reshape(x.shape[0], -1))
-                .cpu()
-                .numpy()
-            )
+            return torch.tensor(self.vqvae.decode(x.numpy()))
 
     @property
-    def bos_id(self) -> int:
-        return self.to_id(self.BOS)
-
-    @property
-    def eos_id(self) -> int:
-        return self.to_id(self.EOS)
-
-    @property
-    def tokens(self) -> set[TokenType]:
-        return set(self.token_to_id.keys())
-
-    def encode(
-        self, images: np.ndarray, drop_eos: bool, drop_bos: bool
-    ) -> list[list[int]]:
-        if not isinstance(images, np.ndarray):
-            raise ValueError(
-                f"Found '{type(images)}' for `image` instead of np.ndarray."
-            )
-        if len(images.shape) == 3:
-            images = images[np.newaxis, ...]
-        sequences = []
-        for batch in batched(images, self.batch_size):
-            sequences.append(self.image_to_sequence(np.array(batch)))
-        seq = np.concatenate(sequences, axis=0)
-        map_func = np.vectorize(lambda word: self.token_to_id[word])
-        seq_tokenized = map_func(seq)
-        if not drop_bos or not drop_eos:
-            rows = seq.shape[0]
-            seq_parts = [seq]
-            if not drop_bos:
-                prefix = np.full((rows, 1), self.bos_id)
-                seq_parts = [prefix, *seq_parts]
-            if not drop_eos:
-                suffix = np.full((rows, 1), self.eos_id)
-                seq_parts = [*seq_parts, suffix]
-            return np.concatenate(seq_parts, axis=1).tolist()
-        else:
-            return seq_tokenized.tolist()
-
-    def decode(self, sequence: list[int]) -> np.ndarray:
-        if not isinstance(sequence, list):
-            raise ValueError(
-                f"Found '{type(sequence)}' for `sequence` instead of list[int]."
-            )
-        sequence_2d = np.array(sequence).reshape(1, 7, 7)
-        with torch.no_grad():
-            return self.vqvae.decode(sequence_2d)
-
-    def decode_batch(self, sequences: Iterator[list[int]]) -> Iterator[np.ndarray]:
-        if not isinstance(sequences, Iterator):
-            raise ValueError(
-                f"Found '{type(sequences)}' for `sequences` instead of "
-                f"Iterator[list[int]]."
-            )
-        for t in sequences:
-            yield self.decode(t)
+    def codebook_size(self) -> int:
+        return self.vqvae.n_embeddings
 
 
 def generate_samples_colored_mnist_image(
     *,
     seq_len: int,
-    tokenizer: VQVAEColoredMNISTImageTokenizer,
+    tokenizer: VQVAEImageTokenizer,
     transformer: GPT2Decoder,
     decode: bool,
 ) -> torch.Tensor:
@@ -165,8 +81,8 @@ class ColoredMNISTImageAndTextTokenizer(BaseTokenizer):
         self.device = get_device()
 
         text_tokens = get_unique_tokens()
-        self.image_tokenizer = VQVAEColoredMNISTImageTokenizer(
-            token_id_offset=token_id_offset
+        self.image_tokenizer = VQVAEImageTokenizer(
+            vqvae=VQVAEMNIST(), token_id_offset=token_id_offset
         )
         self.text_tokenizer = WordTokenizer(
             tokens=text_tokens,
@@ -484,7 +400,7 @@ def sample_completion_image(
     images: np.ndarray,
     show_fraction: float,
     num_completions: int,
-    tokenizer: VQVAEColoredMNISTImageTokenizer,
+    tokenizer: VQVAEImageTokenizer,
     model: GPT2Decoder,
     skip_indices: set[int] | None = None,
 ) -> list[np.ndarray]:
